- Uso de jwt - Sistema de autenticación

+++ src/controllers/ProjectController.ts
@@ -4,7 +4,12 @@ import Project from "../models/Project"
 export class ProjectController {
 
     static createProject = async (req: Request, res: Response) => {
+
         const project = new Project(req.body)
+
+        // Asigna un manager
+        project.manager = req.user.id
+
         try {
             await project.save()
             res.send('Creado Correctamente')
@@ -16,7 +21,11 @@ export class ProjectController {
 
     static getAllProjects = async (req: Request, res: Response) => {
         try {
-            const projects = await Project.find({})
+            const projects = await Project.find({
+                $or: [
+                    { manager: { $in: req.user.id } }
+                ]
+            })
             res.json(projects)
         } catch (error) {
             console.log(error)
@@ -35,6 +44,10 @@ export class ProjectController {
                 return res.status(404).json({ error: error.message })
             }
 
+            if (project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Acción no válida')
+                return res.status(404).json({ error: error.message })
+            }
             res.json(project)
         } catch (error) {
             console.log(error)
@@ -47,6 +60,11 @@ export class ProjectController {
             req.project.projectName = req.body.projectName
             req.project.description = req.body.description
 
+            if (req.project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Solo el Manager puede actualizar un proyecto')
+                return res.status(404).json({ error: error.message })
+            }
+
             await req.project.save()
             res.json('Proyecto Actualizado')
         } catch (error) {
@@ -55,6 +73,11 @@ export class ProjectController {
     }
 
     static deleteProject = async (req: Request, res: Response) => {
+
+        if (req.project.manager.toString() !== req.user.id.toString()) {
+            const error = new Error('Solo el Manager puede actualizar un proyecto')
+            return res.status(404).json({ error: error.message })
+        }
 
         try {
             await req.project.deleteOne()
